PROJ-43 Relative Markerposition wurde nicht richtig in firebase gespeichert

+++ LightsCatcher/app/src/main/java/com/hs_augsburg_example/lightscatcher/dataModels/LightPosition.java
@@ -11,14 +11,14 @@ public class LightPosition {
 
     public int isMostRelevant;
     public int phase;
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
     public LightPosition() {
         // Default constructor required for calls to DataSnapshot.getValue(User.class)
     }
 
-    public LightPosition(int x, int y, int phase, boolean isMostRelevant) {
+    public LightPosition(double x, double y, int phase, boolean isMostRelevant) {
         this.x = x;
         this.y = y;
         this.phase = phase;
